+++ docs/src/conf.py
@@ -24,7 +24,7 @@ copyright = "2024, IDS Srl"
 author = "Emanuele Lena <emanuele.lena@designcoaching.net>"
 
 # The full version, including alpha/beta/rc tags.
-version = "0.2.3"
+version = "0.2.4"
 
 # -- General configuration ------------------------------------------------